Fixes grouping/coloring/sizing by attribute
The option of having multiple tooltips for one time point messed up the
grouping by attribute. This fixes it such that tooltips with multiple tables in
their tooltip end up in their own group "multi". This commit also fixes the
problem of attributes not being displayed in the tooltip of colorby/splitby.