Change dropshadow to look nicer in firefox
Somehow the dropshadow effect is square in the upper left corner. This changes
the blur to be less intense so it looks better in firefox. Also updates the
screenshot. Notice that the svg filter is not implemented in phantomjs so the
test is unfortunately not an accurate representation of the timeline.